Server Api code/Fastapi/app/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
black==23.1a1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.1.0
fastapi==0.87.0
fastjsonschema==2.16.2
filetype==1.2.0
Flask==2.2.2
Flask-SocketIO==5.3.2
flatbuffers==22.10.26
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
google-auth==1.4.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
humanfriendly==10.0
idna==2.8
ImageHash==4.3.1
imageio==2.25.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.15.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lab==7.2
libclang==14.0.6
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==0.4.3
nb-black==1.0.7
nbclassic==0.4.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==3.0
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.4
numpy==1.23.5
nvidia-pyindex==1.0.9
oauthlib==3.2.2
onnxruntime==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
pooch==1.6.0
portpicker==1.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.13.0
PyMatting==1.1.8
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyreadline3==3.4.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.2
pytz==2022.2.1
PyWavelets==1.4.1
pywin32==304
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rembg==2.0.30
requests==2.21.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.9.3
selenium==4.6.0
Send2Trash==1.8.0
simplegeneric==0.8.1
simplejson==3.18.1
six==1.12.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.1
starlette==0.21.0
sympy==1.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.18.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2023.1.23.1
tinycss2==1.1.1
tomli==2.0.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchsummary==1.5.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
txt2tags==3.7
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.24.3
uvicorn==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
